+++ feature/gallery-test/src/test/java/br/com/orcinus/orca/feature/gallery/test/ui/SemanticsNodeInteractionsProviderExtensionsTests.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2023-2024 Orcinus
+ * Copyright © 2023–2024 Orcinus
  *
  * This program is free software: you can redistribute it and/or modify it under the terms of the
  * GNU General Public License as published by the Free Software Foundation, either version 3 of the
@@ -15,51 +15,30 @@
 
 package br.com.orcinus.orca.feature.gallery.test.ui
 
-import androidx.compose.foundation.Image
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.test.assertIsDisplayed
-import androidx.compose.ui.test.filterToOne
 import androidx.compose.ui.test.junit4.createComposeRule
-import androidx.compose.ui.test.onChildren
-import br.com.orcinus.orca.feature.gallery.test.ui.page.isPage
+import br.com.orcinus.orca.feature.gallery.ui.Actions
 import br.com.orcinus.orca.feature.gallery.ui.Gallery
 import br.com.orcinus.orca.platform.autos.theme.AutosTheme
 import org.junit.Rule
 import org.junit.Test
+import org.junit.runner.RunWith
+import org.robolectric.RobolectricTestRunner
 
-internal class SemanticsMatcherExtensionsTests {
+@RunWith(RobolectricTestRunner::class)
+internal class SemanticsNodeInteractionsProviderExtensionsTests {
   @get:Rule val composeRule = createComposeRule()
 
   @Test
-  fun matchesImage() {
+  fun findsCloseActionButton() {
     composeRule
-      .apply {
-        setContent {
-          Image(
-            painterResource(br.com.orcinus.orca.std.image.compose.R.drawable.image),
-            contentDescription = "Image"
-          )
-        }
-      }
-      .onNode(isImage())
+      .apply { setContent { AutosTheme { Actions() } } }
+      .onCloseActionButton()
       .assertIsDisplayed()
   }
 
   @Test
-  fun matchesPager() {
-    composeRule
-      .apply { setContent { AutosTheme { Gallery() } } }
-      .onNode(isPager())
-      .assertIsDisplayed()
-  }
-
-  @Test
-  fun matchesPage() {
-    composeRule
-      .apply { setContent { AutosTheme { Gallery() } } }
-      .onPager()
-      .onChildren()
-      .filterToOne(isPage())
-      .assertIsDisplayed()
+  fun findsPager() {
+    composeRule.apply { setContent { AutosTheme { Gallery() } } }.onPager().assertIsDisplayed()
   }
 }
